+++ tests/test_basic_app.py
@@ -1,8 +1,8 @@
 import pytest
 from flask import Flask
 
-from flask_caching import Cache
-from flask_caching.backends.simplecache import SimpleCache
+from flask_caching_json import Cache
+from flask_caching_json.backends.simplecache import SimpleCache
 
 try:
     import redis  # noqa
@@ -34,7 +34,7 @@ def test_dict_config(app):
 def test_dict_config_initapp(app):
     cache = Cache()
     cache.init_app(app, config={"CACHE_TYPE": "simple"})
-    from flask_caching.backends.simplecache import SimpleCache
+    from flask_caching_json.backends.simplecache import SimpleCache
 
     assert isinstance(app.extensions["cache"][cache], SimpleCache)
 
@@ -42,7 +42,7 @@ def test_dict_config_initapp(app):
 def test_dict_config_both(app):
     cache = Cache(config={"CACHE_TYPE": "null"})
     cache.init_app(app, config={"CACHE_TYPE": "simple"})
-    from flask_caching.backends.simplecache import SimpleCache
+    from flask_caching_json.backends.simplecache import SimpleCache
 
     assert isinstance(app.extensions["cache"][cache], SimpleCache)
 
@@ -81,7 +81,7 @@ def test_app_redis_cache_backend_url_default_db(app, redis_server):
     }
     cache = Cache()
     cache.init_app(app, config=config)
-    from flask_caching.backends.rediscache import RedisCache
+    from flask_caching_json.backends.rediscache import RedisCache
 
     assert isinstance(app.extensions["cache"][cache], RedisCache)
     rconn = app.extensions["cache"][cache]._write_client.connection_pool.get_connection(